# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AppCloud API",
    "description": "Knowledge graph platform for infrastructure dependency mapping and blast-radius analysis.",
    "version": "1.1.0"
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/components/": {
      "get": {
        "summary": "List Components",
        "tags": [
          "Components"
        ],
        "description": "Returns every Component with its containing Application (when one exists). Filterable via the `type` query (api / db / worker / ui).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "type",
            "required": false,
            "description": "Optional filter on Component type"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true,
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string",
                        "example": "payment-api"
                      },
                      "type": {
                        "type": "string",
                        "example": "api",
                        "description": "service / api / worker / database / queue / cache / frontend / platform"
                      },
                      "runtime": {
                        "type": [
                          "null",
                          "string"
                        ],
                        "example": "nodejs"
                      },
                      "appId": {
                        "type": [
                          "null",
                          "string"
                        ],
                        "description": "Containing Application id (or null if unattached)"
                      },
                      "appName": {
                        "type": [
                          "null",
                          "string"
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a Component",
        "tags": [
          "Components"
        ],
        "description": "Creates a Component node, optionally attaching it to an existing Application via `:CONTAINS`. Either `applicationId` or `appId` works (legacy alias).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "additionalProperties": true,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "type": {
                    "type": "string"
                  },
                  "runtime": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "applicationId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Optional parent Application id"
                  },
                  "appId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Alias for applicationId"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string",
                      "example": "payment-api"
                    },
                    "type": {
                      "type": "string",
                      "example": "api",
                      "description": "service / api / worker / database / queue / cache / frontend / platform"
                    },
                    "runtime": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "example": "nodejs"
                    },
                    "appId": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "description": "Containing Application id (or null if unattached)"
                    },
                    "appName": {
                      "type": [
                        "null",
                        "string"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/components/{id}": {
      "get": {
        "summary": "Get one Component",
        "tags": [
          "Components"
        ],
        "description": "Returns the Component, its parent Application (or null), and the Infra resources it owns via `:CONNECTS_TO {via:\"component-mapping\"}`.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Resource UUID or unique name"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string",
                      "example": "payment-api"
                    },
                    "type": {
                      "type": "string",
                      "example": "api",
                      "description": "service / api / worker / database / queue / cache / frontend / platform"
                    },
                    "runtime": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "example": "nodejs"
                    },
                    "appId": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "description": "Containing Application id (or null if unattached)"
                    },
                    "appName": {
                      "type": [
                        "null",
                        "string"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a Component",
        "tags": [
          "Components"
        ],
        "description": "Partial update on `name` / `type` / `runtime`. Other fields preserved.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "runtime": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Resource UUID or unique name"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string",
                      "example": "payment-api"
                    },
                    "type": {
                      "type": "string",
                      "example": "api",
                      "description": "service / api / worker / database / queue / cache / frontend / platform"
                    },
                    "runtime": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "example": "nodejs"
                    },
                    "appId": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "description": "Containing Application id (or null if unattached)"
                    },
                    "appName": {
                      "type": [
                        "null",
                        "string"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a Component",
        "tags": [
          "Components"
        ],
        "description": "DETACH-delete: removes every relationship the Component participates in (including ownership and Component↔Component connections).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Resource UUID or unique name"
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/components/{id}/connections": {
      "post": {
        "summary": "Connect this Component to another (Component → Component edge)",
        "tags": [
          "Components"
        ],
        "description": "Idempotently MERGEs a `:CONNECTS_TO {protocol, port}` edge from this Component to the target. No `via` is set — the endpoint labels (`:Component`→`:Component`) distinguish it from infra-side edges.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetId"
                ],
                "additionalProperties": true,
                "properties": {
                  "targetId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "The other Component to connect to"
                  },
                  "protocol": {
                    "type": "string",
                    "example": "https"
                  },
                  "port": {
                    "type": [
                      "integer",
                      "string",
                      "null"
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Resource UUID or unique name"
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "connected": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/components/{id}/deploy": {
      "post": {
        "summary": "Deploy a Component onto an Infra (manual ownership)",
        "tags": [
          "Components"
        ],
        "description": "Idempotently MERGEs a `:CONNECTS_TO {via:\"component-mapping\", source:\"manual-deploy\", confidence:100}` edge from the Component to the Infra. Use this to record human-confirmed ownership.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "infraId"
                ],
                "additionalProperties": true,
                "properties": {
                  "infraId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Infra node to MERGE a :CONNECTS_TO {via:component-mapping} edge to"
                  }
                }
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Resource UUID or unique name"
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deployed": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "http://appcloud.local"
    }
  ],
  "security": [
    {
      "ApiKey": []
    }
  ]
}